Allow transposition table sizes bigger than 32 Gb (TCEC preparation) (#115)
* support hash size bigger than 32 Gb

* using native march for unix gcc builds

+++ src/tt.h
@@ -91,7 +91,7 @@ public:
 
 private:
     TEntry * m_hash;
-    int m_hashSize;
+    size_t m_hashSize;
     unsigned int m_hashAge;
 };
